Bootstrap
Taking away various elements of CSS and replacing with bootstrap code. mainly did the Resume portion of the site, worked the tables to be responsive to the screen size, and for the main page, used bootstrap framework to place the picture and greeting text next to each other in a more fluid way that would be responsive to the changing screen sizes.