Content reveiw
Decided to scrap the table idea in lieu of a card to hold the idea for projects. as the amount of projects grow, this will help display them in a neat and fashioned way with a picture as a preview. this will also all them to be a bit more readable.

also, made sure the navbar was consistent across pages.